Crate pallet_collective_content

Managed Collective Content Pallet

The pallet provides the functionality to store different types of content. This would typically be used by an on-chain collective, such as the Polkadot Alliance or Ambassador Program.

The pallet stores content as an OpaqueCid, which should correspond to some off-chain hosting service, such as IPFS, and contain any type of data. Each type of content has its own origin from which it can be managed. The origins are configurable in the runtime. Storing content does not require a deposit, as it is expected to be managed by a trusted collective.

Content types:

Collective charter: A single document (OpaqueCid) managed by CharterOrigin.

Collective announcements: A list of announcements managed by AnnouncementOrigin.

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

The pallet weight info trait and its unit implementation.

Type Aliases

OpaqueCid

IPFS compatible CID.